Recommended changes to buttons.
I moved the bottom button before the script tag so it would show up.

I didn't change the other buttons, but the JS that you want to work on the "Click" button is actually activated on the navbar toggle button because that's the first button in the document. I added a suggestion for how to change the JS to work on the "click" button instead.