Recommend using <p> in place of some <h4> or <h5>
Most of the text here is formatted as headings. It think it should instead be that most is formatted as paragraphs, with a few headings to stand out. I think, but I'm not sure, that screen readers and other technology might expect paragraphs where some of the smaller headings are (e.g. where I put comments). Using <p> instead of <h5> and <h4> would mean that you wouldn't have to add tags like <b> to differentiate between a section heading (like Vaccinate) and the contents below it.